Split test_api.py per data type
